+++ packages/block-library/src/separator/index.js
@@ -11,7 +11,7 @@ export const settings = {
 
 	description: __( 'Insert a horizontal line where you want to create a break between ideas.' ),
 
-	icon: <svg role="img" aria-hidden="true" focusable="false" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="none" d="M0 0h24v24H0V0z" /><path d="M19 13H5v-2h14v2z" /></svg>,
+	icon: <svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="none" d="M0 0h24v24H0V0z" /><path d="M19 13H5v-2h14v2z" /></svg>,
 
 	category: 'layout',