+++ packages/editor/src/components/block-list/block-draggable.js
@@ -8,24 +8,39 @@ import classnames from 'classnames';
  */
 import { Draggable } from '@wordpress/components';
 
-function BlockDraggable( { rootClientId, index, clientId, layout, isDragging, ...props } ) {
+const BlockDraggable = ( { clientId, rootClientId, blockElementId, layout, order, isDragging, onDragStart, onDragEnd } ) => {
 	const className = classnames( 'editor-block-list__block-draggable', {
 		'is-visible': isDragging,
 	} );
 
 	const transferData = {
 		type: 'block',
-		fromIndex: index,
+		fromIndex: order,
 		rootClientId,
 		clientId,
 		layout,
 	};
 
 	return (
-		<Draggable className={ className } transferData={ transferData } { ...props }>
-			<div className="editor-block-list__block-draggable-inner"></div>
+		<Draggable
+			elementId={ blockElementId }
+			transferData={ transferData }
+			onDragStart={ onDragStart }
+			onDragEnd={ onDragEnd }
+		>
+			{
+				( { onDraggableStart, onDraggableEnd } ) => (
+					<div
+						className={ className }
+						onDragStart={ onDraggableStart }
+						onDragEnd={ onDraggableEnd }
+						draggable
+					>
+						<div className="editor-block-list__block-draggable-inner"></div>
+					</div> )
+			}
 		</Draggable>
 	);
-}
+};
 
 export default BlockDraggable;