+++ packages/data/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v2.0.0 (Unreleased)
 
-- Breaking: The `withRehdyration` function is removed. Use the persistence plugin instead.
-- Breaking: The `loadAndPersist` function is removed. Use the persistence plugin instead.
+### Breaking Change
+
+- The `withRehdyration` function is removed. Use the persistence plugin instead.
+- The `loadAndPersist` function is removed. Use the persistence plugin instead.
+- `restrictPersistence`, `setPersistenceStorage` and  `setupPersistence` functions have been removed. Please use the data persistence plugin instead.